+++ packages/web3-utils/CHANGELOG.md
@@ -174,7 +174,8 @@ Documentation:
 ### Added
 
 -   As a replacement of the node EventEmitter, a custom `EventEmitter` has been implemented and exported. (#6398)
--   `SocketProvider` now contains public function `getPendingRequestQueueSize` (#6451)
+-   `SocketProvider` now contains public function `getPendingRequestQueueSize`, `getSentRequestsQueueSize` (#6451)
+-   Added `safeDisconnect` as a `SocketProvider` method to disconnect only when request queue size and send request queue size is 0 (#6451) 
 
 ### Fixed